Move varuint functions into plain_text frame_helper (#587)
