+++ aioesphomeapi/_frame_helper/plain_text.pxd
@@ -7,6 +7,11 @@ cdef bint TYPE_CHECKING
 cdef object WRITE_EXCEPTIONS
 cdef object bytes_to_varuint, varuint_to_bytes
 
+cpdef _varuint_to_bytes(cython.int value)
+
+@cython.locals(result=cython.int, bitpos=cython.int, val=cython.int)
+cpdef _bytes_to_varuint(cython.bytes value)
+
 cdef class APIPlaintextFrameHelper(APIFrameHelper):
 
     @cython.locals(
