+++ tests/test__frame_helper.py
@@ -4,12 +4,19 @@ import pytest
 
 from aioesphomeapi._frame_helper import APINoiseFrameHelper, APIPlaintextFrameHelper
 from aioesphomeapi._frame_helper.base import WRITE_EXCEPTIONS
+from aioesphomeapi._frame_helper.plain_text import _bytes_to_varuint as bytes_to_varuint
+from aioesphomeapi._frame_helper.plain_text import (
+    _cached_bytes_to_varuint as cached_bytes_to_varuint,
+)
+from aioesphomeapi._frame_helper.plain_text import (
+    _cached_varuint_to_bytes as cached_varuint_to_bytes,
+)
+from aioesphomeapi._frame_helper.plain_text import _varuint_to_bytes as varuint_to_bytes
 from aioesphomeapi.core import (
     BadNameAPIError,
     InvalidEncryptionKeyAPIError,
     SocketAPIError,
 )
-from aioesphomeapi.util import varuint_to_bytes
 
 PREAMBLE = b"\x00"
 
@@ -234,3 +241,25 @@ async def test_noise_incorrect_name():
 
     with pytest.raises(BadNameAPIError):
         await helper.perform_handshake(30)
+
+
+VARUINT_TESTCASES = [
+    (0, b"\x00"),
+    (42, b"\x2a"),
+    (127, b"\x7f"),
+    (128, b"\x80\x01"),
+    (300, b"\xac\x02"),
+    (65536, b"\x80\x80\x04"),
+]
+
+
+@pytest.mark.parametrize("val, encoded", VARUINT_TESTCASES)
+def test_varuint_to_bytes(val, encoded):
+    assert varuint_to_bytes(val) == encoded
+    assert cached_varuint_to_bytes(val) == encoded
+
+
+@pytest.mark.parametrize("val, encoded", VARUINT_TESTCASES)
+def test_bytes_to_varuint(val, encoded):
+    assert bytes_to_varuint(encoded) == val
+    assert cached_bytes_to_varuint(encoded) == val
